chore: Move StatusBadge from design to turboui
