+++ turboui/src/index.tsx
@@ -3,3 +3,4 @@ export * from "./TestableElement";
 export * from "./Button";
 export * from "./Link";
 export * from "./Menu";
+export * from "./StatusBadge";
